python-app/requirements.txt
streamlit
openai
PyPDF2
python-docx
transformers
torch
pymongo